Moved util package: kafe -> kafe/fit
