+++ kafe/fit/__init__.py
@@ -133,6 +133,7 @@ from ._base.profile import *   # just for `ContoursProfiler` -> better way?
 from .indexed import *
 from .histogram import *
 from .tools import *
+from .util import *
 from .xy import *
 from .xy_multi import *
